conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=0_gnu
absl-py=0.9.0=py38h32f6830_1
astunparse=1.6.3=pypi_0
attrs=19.3.0=py_0
backcall=0.1.0=py_0
blas=1.0=mkl
bleach=3.1.5=pyh9f0ad1d_0
blinker=1.4=py_1
bokeh=2.0.2=pypi_0
brotlipy=0.7.0=py38h1e0a361_1000
bzip2=1.0.8=h516909a_2
c-ares=1.15.0=h516909a_1001
ca-certificates=2020.4.5.1=hecc5488_0
cachetools=3.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.4.5.1=py38h32f6830_0
cffi=1.14.0=py38hd463f26_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.4.1=py_0
colour=0.1.5=pypi_0
cryptography=2.9.2=py38h766eaa4_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
cytoolz=0.10.1=py38h516909a_0
dask-core=2.16.0=py_0
datetime=4.3=pypi_0
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.1.1=pypi_0
distributed=2.16.0=pypi_0
entrypoints=0.3=py38h32f6830_1001
expat=2.2.9=he1b5a44_2
ffmpeg=4.2=h167e202_0
flask=1.0.3=pypi_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=he06d7ca_0
fsspec=0.7.4=pypi_0
future=0.18.2=pypi_0
geojson=2.5.0=pypi_0
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.2.1=h516909a_2
glib=2.64.2=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.5=hd3a4fd2_1002
google-auth=1.14.3=pyh9f0ad1d_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=he1b5a44_1001
grpcio=1.27.2=py38hf8bcb03_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_100
heapdict=1.0.1=pypi_0
icu=64.2=he1b5a44_1
idna=2.9=py_1
imagecodecs-lite=2019.12.3=py38h1e0a361_0
imageio=2.8.0=py_0
importlib-metadata=1.6.0=py38h32f6830_0
importlib_metadata=1.6.0=0
intel-openmp=2020.1=217
ipykernel=5.2.1=py38h23f93f0_0
ipython=7.14.0=py38h32f6830_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=pypi_0
itsdangerous=1.1.0=pypi_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.2=py38_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.15.1=py_0
jpeg=9c=h14c3975_1001
json5=0.9.0=py_0
jsonpatch=1.25=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py38h32f6830_1
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-full-width=1.2.0=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter_client=6.1.3=py_0
jupyter_contrib_core=0.3.3=py_2
jupyter_core=4.6.3=py38h32f6830_1
jupyter_nbextensions_configurator=0.4.1=py38_0
jupyterlab=2.1.2=py_0
jupyterlab_code_formatter=1.3.1=py_0
jupyterlab_server=1.1.3=py_0
kiwisolver=1.2.0=py38hbf85e49_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libclang=9.0.1=default_hde54327_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.2.0=h24d8f2e_2
libiconv=1.15=h516909a_1006
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.2.0=py38_6
libpng=1.6.37=hed695b0_1
libprotobuf=3.11.4=h8b12597_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc3755c2_3
libuuid=2.32.1=h14c3975_1000
libuv=1.34.0=h516909a_0
libwebp=1.0.2=h56121f0_5
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
llvmlite=0.32.1=pypi_0
locket=0.2.0=pypi_0
lxml=4.5.0=pypi_0
lz4-c=1.9.2=he1b5a44_1
markdown=3.2.2=py_0
markupsafe=1.1.1=py38h1e0a361_1
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py38h2af1d28_0
mistune=0.8.4=py38h1e0a361_1001
mkl=2019.4=243
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
mpld3=0.3=pypi_0
msgpack=1.0.0=pypi_0
multiprocess=0.70.9=pypi_0
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.0.6=py_0
nbresuse=0.3.3=py_0
ncurses=6.1=hf484d3e_1002
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_1
ninja=1.9.0=py38hfd86e86_0
nodejs=13.13.0=hf5d1a2b_0
notebook=6.0.3=py38h32f6830_0
nspr=4.25=he1b5a44_0
nss=3.47=he751ad9_0
numba=0.49.1=pypi_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
oauthlib=3.0.1=py_0
olefile=0.46=py_0
opencv=4.2.0=py38_6
opencv-python=4.2.0.34=pypi_0
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1g=h516909a_0
packaging=20.1=py_0
pandarallel=1.4.8=pypi_0
pandas=1.0.3=py38h0573a6f_0
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py_1
parso=0.7.0=pyh9f0ad1d_0
partd=1.1.0=pypi_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.1.2=py38h9776b28_0
pip=20.1=pyh9f0ad1d_0
pixiedust=1.1.18=pypi_0
pixman=0.38.0=h516909a_1003
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.5=py_0
protobuf=3.11.4=py38h950e882_1
psutil=5.7.0=py38h1e0a361_1
pthread-stubs=0.4=h14c3975_1001
ptvsd=4.3.2=py38h516909a_1
ptyprocess=0.6.0=py_1001
py-opencv=4.2.0=py38h23f93f0_6
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38ha8c2ead_3
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.16.0=py38h1e0a361_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.2=he5300dc_7_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.0=py3.8_cuda10.2.89_cudnn7.6.5_0
pytorch-gradcam=0.2.1=pypi_0
pytz=2020.1=py_0
pywavelets=1.1.1=py38h8790de6_1
pyyaml=5.3.1=py38h1e0a361_0
pyzmq=19.0.1=py38ha71036d_0
qt=5.12.5=hd8c4c69_1
readline=8.0=hf8c457e_0
requests=2.23.0=pyh8c360ce_2
requests-oauthlib=1.2.0=py_0
rsa=4.0=py_0
scikit-image=0.17.2=py38hcb8c335_0
scikit-learn=0.23.1=py38h3a94b23_0
scipy=1.4.1=py38h0b6359f_0
seaborn=0.10.1=pypi_0
send2trash=1.5.0=py_0
setuptools=46.3.0=py38h32f6830_0
shap=0.35.0=py38hcb8c335_0
six=1.14.0=py_1
sortedcontainers=2.1.0=pypi_0
sqlite=3.30.1=hcee41ef_0
swifter=0.304=pypi_0
tblib=1.6.0=pypi_0
tensorboard=2.1.1=py_1
tensorboardx=2.0=py_0
terminado=0.8.3=py38h32f6830_1
testpath=0.4.4=py_0
threadpoolctl=2.0.0=pyh5ca1d4c_0
tifffile=2020.5.11=py_0
tk=8.6.10=hed695b0_0
toolz=0.10.0=py_0
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.6.0=py38_cu102
tornado=6.0.4=py38h1e0a361_1
tqdm=4.46.0=pyh9f0ad1d_0
traitlets=4.3.3=py38h32f6830_1
typing-extensions=3.7.4.2=pypi_0
urllib3=1.25.9=py_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.9=pyh9f0ad1d_0
webencodings=0.5.1=py_1
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py_1
widgetsnbextension=3.5.1=pypi_0
x264=1!152.20180806=h14c3975_0
xeus=0.23.14=h4d8c418_0
xeus-python=0.7.1=py38hbf85e49_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h516909a_0
yaml=0.2.4=h516909a_0
zeromq=4.3.2=he1b5a44_2
zict=2.0.0=pypi_0
zipp=3.1.0=py_0
zlib=1.2.11=h516909a_1006
zope-interface=5.1.0=pypi_0
zstd=1.4.4=h6597ccf_3
